added local json, split QuizQuestion component, new styles
