+++ src/components/Quiz/Quiz.css
@@ -33,11 +33,15 @@
 }
 
 #gradbox {
-    height: 20px;
+    height: 30px;
     background-color: #56b004; 
     background-image: linear-gradient(to right, #ff9500 0%, rgba(0,0,0,0) 0%);
     border-radius: 3px;
     margin: 10px 0px;
+    border: 3px solid #fff;
+    display: flex;
+    justify-content: center;
+    padding: 2px;
 }
 
 .main {
